Dev合并至beta

See merge request yunya_platform/yunya-dental!4

+++ yunya-report/yunya-middle-table/src/main/java/com/yunya/middletable/service/patient/BasePatientBiz.java
@@ -1,7 +1,9 @@
 package com.yunya.middletable.service.patient;
 
+import com.google.common.collect.Lists;
 import com.yunya.feign.report.domain.form.PullForm;
 import com.yunya.feign.report.domain.model.MessageModel;
+import com.yunya.feign.report.domain.vo.PatientTreatInfoVo;
 import com.yunya.framework.common.biz.BaseBiz;
 import com.yunya.framework.common.utils.StringHelper;
 import com.yunya.middletable.dao.patient.PatientBaseInfoMapper;
@@ -17,17 +19,24 @@ import com.yunya.models.report.BasePatient;
 import com.yunya.models.report.BasePatientMember;
 import lombok.extern.slf4j.Slf4j;
 import org.joda.time.DateTime;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import tk.mybatis.mapper.entity.Example;
 
 import javax.annotation.Resource;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
+import java.util.function.Function;
+
+import static java.util.stream.Collectors.toMap;
+
 
 /**
  * 简介: 报表服务患者信息同步
@@ -42,15 +51,15 @@ import java.util.concurrent.Future;
 @Transactional(rollbackFor = Exception.class)
 public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
   /** 注入对象 */
-  @Autowired private PatientBaseInfoMapper patientBaseInfoMapper;
+  @Resource private PatientBaseInfoMapper patientBaseInfoMapper;
 
-  @Autowired private PatientOriginMapper patientOriginMapper;
+  @Resource private PatientOriginMapper patientOriginMapper;
 
-  @Autowired private BasePatientMemberBiz basePatientMemberBiz;
+  @Resource private BasePatientMemberBiz basePatientMemberBiz;
 
-  @Autowired private PatientPrepaymentsInfoMapper patientPrepaymentsInfoMapper;
+  @Resource private PatientPrepaymentsInfoMapper patientPrepaymentsInfoMapper;
 
-  @Autowired private BasePatientMemberMapper basePatientMemberMapper;
+  @Resource private BasePatientMemberMapper basePatientMemberMapper;
 
   /** 多线程 */
   @Resource(name = "customizeThreadPool")
@@ -72,7 +81,8 @@ public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
         addPrepaidInfo(patient);
         break;
       case 1:
-        mapper.updateByPrimaryKeySelective(patient);
+        assert patient != null;
+        mapper.updateByPrimaryKeySelective(getPatientInfo(patient));
         break;
       case 2:
         PatientBaseInfo patientBaseInfo = patientBaseInfoMapper.selectByPrimaryKey(patientId);
@@ -88,20 +98,63 @@ public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
   }
 
   /**
+   * 获取患者初末诊日期
+   * @param patient 患者信息
+   * @return 患者信息
+   */
+  private BasePatient getPatientInfo(BasePatient patient) {
+    if (patient != null){
+      // 初诊信息
+      PatientTreatInfoVo firstTreatInfo = mapper.selectFirstVisitInfo(patient.getPatientId());
+      if (null != firstTreatInfo) {
+        patient.setFirstVisitDate(getDateTime(firstTreatInfo.getTreatDate()));
+        patient.setFirstVisitDoctors(firstTreatInfo.getTreatDentistName());
+        patient.setFirstVisitOutpatient(firstTreatInfo.getTreatOutpatient());
+      }
+      // 末诊信息
+      PatientTreatInfoVo lastTreatInfo = mapper.selectLastVisitInfo(patient.getPatientId());
+      if (null != lastTreatInfo) {
+        patient.setLastVisitDate(getDateTime(lastTreatInfo.getTreatDate()));
+        patient.setLastVisitDoctors(lastTreatInfo.getTreatDentistName());
+        patient.setLastVisitOutpatient(lastTreatInfo.getTreatOutpatient());
+      }
+      return patient;
+    }
+   return null;
+  }
+
+  public Date getDateTime(String dateStr){
+    Date date = null;
+    //获得SimpleDateFormat类，我们转换为yyyy-MM-dd的时间格式
+    SimpleDateFormat sf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+    try {
+    //使用SimpleDateFormat的parse()方法生成Date
+     date = sf.parse(dateStr);
+    //打印Date
+      System.out.println(date);
+    } catch (ParseException e) {
+      e.printStackTrace();
+    }
+    return date;
+  }
+
+  /**
    * 添加患者信息是添加预付款信息
    *
    * @param patient 患者信息
    */
   public void addPrepaidInfo(BasePatient patient) {
-    PatientPrepaymentsInfo patientPrepaymentsInfo = new PatientPrepaymentsInfo();
-    patientPrepaymentsInfo.setPatientId(patient.getPatientId());
-    PatientPrepaymentsInfo patientPrepayments =
-        patientPrepaymentsInfoMapper.selectOne(patientPrepaymentsInfo);
-    if (patientPrepayments != null) {
-      BasePatientMember basePatientMember =
-          basePatientMemberBiz.getPatientMemberInfo(patientPrepayments.getId(), 1);
-      basePatientMemberMapper.deleteByPrimaryKey(basePatientMember);
-      basePatientMemberMapper.insert(basePatientMember);
+    if (patient != null){
+      PatientPrepaymentsInfo patientPrepaymentsInfo = new PatientPrepaymentsInfo();
+      patientPrepaymentsInfo.setPatientId(patient.getPatientId());
+      PatientPrepaymentsInfo patientPrepayments =
+              patientPrepaymentsInfoMapper.selectOne(patientPrepaymentsInfo);
+      if (patientPrepayments != null) {
+        BasePatientMember basePatientMember =
+                basePatientMemberBiz.getPatientMemberInfo(patientPrepayments.getId(), 1);
+        basePatientMemberMapper.deleteByPrimaryKey(basePatientMember);
+        basePatientMemberMapper.insert(basePatientMember);
+      }
     }
   }
 
@@ -199,7 +252,7 @@ public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
   }
 
   /**
-   * 查询患者质料信息
+   * 查询患者资料信息
    *
    * @param id 患者id
    * @return BasePatient
@@ -207,12 +260,76 @@ public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
   public BasePatient selectPatientInfo(Integer id) {
     BasePatient basePatient = new BasePatient();
     basePatient.setPatientId(id);
-    BasePatient basePatientVO = mapper.selectOne(basePatient);
-    if (basePatient != null) {
-      return basePatientVO;
-    }
-    return null;
+    return mapper.selectOne(basePatient);
   }
 
 
+  /**
+   * 批量更新患者信息
+   */
+  public void updPatientInfo() throws InterruptedException {
+    // 查询有过初诊的患者id
+    List<BasePatient> patientIdList  = mapper.selectPatientIdList();
+    // k 患者id v 初诊信息
+    Map<Integer, PatientTreatInfoVo> firstInfoMap = getFirstInfoMap();
+    // k 患者id v 末诊信息
+    Map<Integer, PatientTreatInfoVo> lastTreatInfoMap = getLastTreatInfoMap();
+
+    if (!StringHelper.isEmpty(patientIdList)){
+     for (BasePatient  basePatient : patientIdList) {
+         // 获取初诊信息
+         PatientTreatInfoVo firstInfo = firstInfoMap.get(basePatient.getPatientId());
+         if (null != firstInfo.getPatientId()){
+             basePatient.setFirstVisitDate(getDateTime(firstInfo.getTreatDate()));
+             basePatient.setFirstVisitDoctors(firstInfo.getTreatDentistName());
+             basePatient.setFirstVisitOutpatient(firstInfo.getTreatOutpatient());
+           }
+         // 获取末诊信息
+         PatientTreatInfoVo lastTreatInfo = lastTreatInfoMap.get(basePatient.getPatientId());
+         if (null != lastTreatInfo.getPatientId()){
+             basePatient.setLastVisitDate(getDateTime(lastTreatInfo.getTreatDate()));
+             basePatient.setLastVisitDoctors(lastTreatInfo.getTreatDentistName());
+             basePatient.setLastVisitOutpatient(lastTreatInfo.getTreatOutpatient());
+           }
+     }
+
+     List<List<BasePatient>> partitionLists = Lists.partition(patientIdList, 100);
+     CountDownLatch countDownLatch = new CountDownLatch(partitionLists.size());
+     long start = System.currentTimeMillis();
+     for (List<BasePatient> basePatientList: partitionLists) {
+       importExcelThreadPool.execute(() ->{
+         try {
+           mapper.updatePatientInfoList(basePatientList);
+           countDownLatch.countDown();
+         } catch (Exception e) {
+           log.info("患者信息批量修改异常",e);
+         }
+       });
+     }
+     countDownLatch.await();
+     long end = System.currentTimeMillis();
+     log.info("患者信息批量修改完成，时长：[{}]秒",(end - start) / 1000);
+   }
+
+  }
+
+  /**
+   * 封装患者初诊信息
+   * @return k 患者id v 初诊信息
+   */
+  private Map<Integer, PatientTreatInfoVo> getFirstInfoMap() {
+    // 初诊信息
+    List<PatientTreatInfoVo> firstInfoVoList = mapper.selectFirstVisitInfoList();
+    return firstInfoVoList.stream().collect(toMap(PatientTreatInfoVo::getPatientId, Function.identity()));
+  }
+
+  /**
+   * 封装患者末诊信息
+   * @return k 患者id v 末诊信息
+   */
+  private Map<Integer, PatientTreatInfoVo> getLastTreatInfoMap() {
+    // 末诊信息
+    List<PatientTreatInfoVo> firstInfoVoList = mapper.selectFirstVisitInfoList();
+    return firstInfoVoList.stream().collect(toMap(PatientTreatInfoVo::getPatientId, Function.identity()));
+  }
 }
